+++ Sources/HealthKitDataSource/HealthKit Extensions/NSPredicate+Sendable.swift
