+++ Sources/HealthKitDataSource/HealthKitDeliveryStartSetting.swift
@@ -1,0 +1,16 @@
+//
+// This source file is part of the CardinalKit open-source project
+//
+// SPDX-FileCopyrightText: 2022 CardinalKit and the project authors (see CONTRIBUTORS.md)
+//
+// SPDX-License-Identifier: MIT
+//
+
+
+/// <#Description#>
+public enum HealthKitDeliveryStartSetting: Equatable {
+    /// <#Description#>
+    case manual
+    /// <#Description#>
+    case applicationWillLaunch
+}
