+++ Tests/UITests/TestApp/Info.plist.license
@@ -1,0 +1,5 @@
+This source file is part of the CardinalKit open-source project
+
+SPDX-FileCopyrightText: 2022 CardinalKit and the project authors (see CONTRIBUTORS.md)
+
+SPDX-License-Identifier: MIT
