+++ Tests/UITests/TestApp/Shared/TestAppDelegate.swift
@@ -18,13 +18,16 @@ class TestAppDelegate: CardinalKitAppDelegate {
     override var configuration: Configuration {
         Configuration(standard: TestAppStandard()) {
             ObservableComponentTestsComponent(message: "Passed")
-//            if HKHealthStore.isHealthDataAvailable() {
-//                HealthKit(
-//                    [
-//                        Collect(type: HKQuantityType.electrocardiogramType())
-//                    ]
-//                )
-//            }
+            if HKHealthStore.isHealthDataAvailable() {
+                HealthKit(
+                    [
+                        Collect(sampleType: HKQuantityType.electrocardiogramType())
+                    ],
+                    adapter: {
+                        TestAppHealthKitAdapter()
+                    }
+                )
+            }
             SecureStorage()
             LocalStorage()
         }
